Move chrono-tz timestamp test to arrow/tests
